Add Debian 7.0.0 definitions
This is a start on having Debian 7.0 basebox support in bento.

Manual steps required for building at this time:

- Hit enter to select American English on "Configure the Keyboard"
- Hit enter to NOT scan another media on "Configure the package manager"
- Hit enter to use network mirror on "Configure the package manager"

64 and 32 bit baseboxes are already uploaded to the S3 bucket built
with these definitions. They include Chef 11.4.4 because we haven't
made baseboxes provisionerless (yet).

The API that serves up Chef omnibus packages doesn't have Debian 7
yet, so these have a chef-client.sh that hits the API with the Debian
6 version requested.