Adding vmtools script to Debian.
Also changing the shutdown command to be more inline with others as it
does not seem to timeout for me anymore especially when I build each box
individually using Packers's -only flag.

I'd missed out the vmtools script and until a concrete decision on
installing Chef is made I matched the other templates by including the
chef_omnibus script.

Also changed the order of the scripts to match the other templates.