Update to FreeBSD 9.2.
This required a lot of work to redo the ZFS install.sh. Under FreeBSD 9.2,
you can no longer mount the temporary zroot under /mnt during the install,
since / is a read-only filesystem (being the CD-ROM).

I took the opportunity to remove a lot of the legacy cruft in install.sh,
like all the gpart.nop hackery. Much was stolen from this procedure:
https://wiki.freebsd.org/RootOnZFS/GPTZFSBoot