Fix mac address issue for ens33 on centos 7
Before I had an issue using the box with a static ip for NFS sharing in
vagrant, resulting in

```
ERROR: [/etc/sysconfig/network-scripts/ifdown-eth] Device ens33 has MAC
address 00:0C:29:F8:96:74, instead of configured address
00:0C:29:59:C4:97. Ignoring.
```

While the cleanup script already fixed this for `eth0`, CentOS7 seems to
have changed the device naming, therefor it's also needed for `ens33`.

Instead of removing the text of the entries, I opted for deleting the
whole line as there's no need for empty lines in the ifcfg-file.