Remove `chef_version` user variable & remove Chef installation option.
In an effort to simplify the potential for behavior proliferation that
user variables allows, the `chef_version` user variable is being
dropped. The default behavior was to *not* install Chef for all Bento
boxes as downstream tooling such as Test Kitchen or vagrant-omnibus can
handle Chef installations at boot time.