Update freebsd-update for 10.2
The FreeBSD 10.2 freebsd-update utility supports a `--not-running-from-cron` argument, so patching it to run without a TTY is no longer required. Fix failed fetching of files caused by empty proxy environment variables.

Note that freebsd-update fetch & install will break if there arent any patches available for the release.

I've tried to handle the install case when no updates are needed by capturing the fetch output and checking the last line. This results in some odd behaviour when there are updates available and causes the install to fail completely. As such I've removed these changes.