Initial commit for adding Fedora 14 to Bento.
The purpose of having Fedora 14 was to build/setup image that's compatible with
Arista EOS [1] which is based on Fedora 14.

Older version of Fedora/RHEL default `sudoers` to `requiretty` which break
Vagrant ability to run shell scripts due to unable to sudo, this is now fixed.

Adding required dkms and other packages to build virtualbox guest properly.

[1] https://eos.arista.com/