[macos-10.12] Allow Packer to pick up memory user var.
For some reason with Pack 0.12.3, the current `"memory"` user variable
is not being picked up and a default value of `1024` is always being
used, leading to macOS 10.12 not being able to finish its installation.
For some reason, changing the user variable from `"memory"` to almost
anything else (in this case, `"memsize"`) allows the user variable to be
picked up.

Note that this was observed with the `vmware-iso` builder and may be the
only affected one. I verified this behavior by running:

```
env PACKER_LOG=1 bundle exec bento build --only=vmware-iso --debug macos-10.12
```

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>