Setup Fedora with a single partition to shrink the image
This seems to allow the compression to work better on the virtual disk. The vbox image goes from 703 megs to 317 megs. I also removed some hardware related package while I was in there and the DNF yum compatibility, which I think still gets pulled in due to other deps, but we don't need to explicitly set that.

Signed-off-by: Tim Smith <tsmith@chef.io>