Make Ubuntu APT use archive.ubuntu.com rather than the US domain
By default the Ubuntu installer sets the repository hostnames in
`/etc/apt/sources.list` to the country-specific mirrors based on the
locale chosen, which for this image is `us.archive.ubuntu.com`.

The US mirror is slow for users outside the US, so a better option is
to use `archive.ubuntu.com`, which automatically resolves to the
closest official country mirror to them. This is also the hostname
used in the official Ubuntu Docker images.

Fixes #810.

Signed-off-by: Ed Morley <edmorley@users.noreply.github.com>