Fix linux-firmware dependency on CentOS 8.2
The kernel-core package requires the linux-firmware package and
excluding it in the kickstart file breaks the installation.

After the system is installed, though, it can be removed successfully.

The iwl* firmware packages are the only ones installed by @core, so
they're the only ones that need explicit exclusion in the kickstart.