Fix CentOS 8.2 build breakage due to authconfig
The auth statement in the kickstart file has been deprecated in favor of
authselect. The build is failing due to missing /usr/sbin/authconfig
even when the authselect-compat package is added to the package list.

In any case, the existing auth directive was specifying things that are
now the defaults, so it can be removed.